+++ src/xcore/config.cpp
@@ -230,6 +230,7 @@ void saveConfig() {
 	fprintf(cfile, "dmsize = %i\n", conf.dbg.dmsize);
 	fprintf(cfile, "stack.offset = %i\n", conf.dbg.stackofs);
 	fprintf(cfile, "scr.zoom = %i\n", conf.dbg.scrzoom);
+	fprintf(cfile, "regs.layout = %i\n", conf.dbg.reglayout);
 	fprintf(cfile, "dim.address = %s\n", YESNO(conf.dbg.dimadr));
 	fprintf(cfile, "dim.opcodes = %s\n", YESNO(conf.dbg.dimops));
 	fprintf(cfile, "syntax.colors = %s\n", YESNO(conf.dbg.synhl));
@@ -433,6 +434,7 @@ void loadConfig() {
 	conf.dbg.dmsize = 127;
 	conf.dbg.stackofs = -2;
 	conf.dbg.scrzoom = 1;
+	conf.dbg.reglayout = DBG_REGS_AUTO;
 	conf.dbg.dimadr = 0;
 	conf.dbg.dimops = 1;
 	conf.dbg.synhl = 1;
@@ -513,6 +515,9 @@ void loadConfig() {
 					}
 					if ((pnam == "scr.zoom") && (arg.i > 0) && (arg.i < 4))
 						conf.dbg.scrzoom = arg.i;
+					if ((pnam == "regs.layout") && ((arg.i == DBG_REGS_AUTO) || (arg.i == DBG_REGS_1COL)
+							|| (arg.i == DBG_REGS_2COL) || (arg.i == DBG_REGS_WIDE)))
+						conf.dbg.reglayout = arg.i;
 					if (pnam == "dim.address") conf.dbg.dimadr = arg.b;
 					if (pnam == "dim.opcodes") conf.dbg.dimops = arg.b;
 					if (pnam == "syntax.colors") conf.dbg.synhl = arg.b;